LI UG: Update autogen documentation
This commit updates the autogenerated .tex files and the hard-coded .tex
files that refer to the autogen files.  The autogen files were created
using a modified version of the parse_ocean_xml_registry.py script in
the branch landice_userguide_autogen_mods in my fork of the MPAS-Tools
repo.  This script has been updated to work with pools.  See notes in
model_configuration.tex about how to use it for the LI UG.